requirements.txt
dj-database-url==0.4.1
Flask==0.12
Django==1.11
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.3.0
requests==2.9.1
telepot
asyncio
python-telegram-bot